% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_binning.R
\name{cyto_binning}
\alias{cyto_binning}
\title{Flow cytometric data binning}
\usage{
cyto_binning(
  gh,
  node = "root",
  channels,
  xlimits,
  ylimits,
  binnumber,
  vectorize = TRUE
)
}
\arguments{
\item{gh}{a GatingHierarchy object.}

\item{node}{name of the node (gate).}

\item{channels}{index of two channels (columns) in the cytometric data matrix, a vector with 2 elements.}

\item{xlimits}{lower and upper limits of the 1st channel, a vector with 2 elements. Must be at least as low/high as the minimun/maximum value within the channel node.}

\item{ylimits}{lower and upper limits of the 2nd channel, a vector with 2 elements. Must be at least as low/high as the minimun/maximum value within the channel node.}

\item{binnumber}{number of bins (compartments) to be created along the channel axes.}

\item{vectorize}{logical indicating whether the resulting binning matrix should be converted into a vector for easier comparison with other GatingHierarchy objects.}
}
\value{
A binning matrix and (optionally) its vectorized form. In the case of both, a list with the matrix as the first and the vector as the second element.
}
\description{
Divide the data ranges of two selected channels of a flow cytometric data set (a GatingHierarchy object) into \code{n} categories.
The process results in a matrix with \code{n} x \code{n} compartments (bins) in the space defined by the selected channels,
assigning each detected cell/event to a sepcific bin according to its relevant channel signals.
}
\examples{
library(flowCore)
library(flowWorkspace)

# loading flow cytometric data of a phytoplankton sample as a GatingSet object
phyto_gs   <- load_gs(path = system.file("extdata", "phyto_flowcyt_sample", package = "cytobins"))
# creating a separate GatingHierarchy object
phyto_gh   <- phyto_gs[[1]]

# data binning along the axes of optical channels FSC and FL3, with 64x64 bins
# performed on a gated population ("chl") of phyto_gh
phyto_bins <- cyto_binning(gh = phyto_gh, node = "chl", channels = c(1,9),
                           xlimits = c(3,6), ylimits = c(3,7), binnumber = 64)
# the generated binning matrix
phyto_bins[[1]]
# the generated binning vector
phyto_bins[[2]]
}
