% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_binning.R
\name{cyto_binning}
\alias{cyto_binning}
\title{Binning flow cytometric data}
\usage{
cyto_binning(
  gh,
  node = "root",
  channels,
  xlimits,
  ylimits,
  binnumber,
  vectorize = TRUE
)
}
\arguments{
\item{gh}{a GatingHierarchy object.}

\item{node}{name of the node (gate).}

\item{channels}{number of the channels (columns) in the cytometric data matrix, a vector with 2 elements.}

\item{xlimits}{lower and upper limits of the 1st channel, a vector with 2 elements.}

\item{ylimits}{lower and upper limits of the 2nd channel, a vector with 2 elements.}

\item{binnumber}{number of bins (compartments) to be created along the channel axes.}

\item{vectorize}{logical indicating whether the resulting binning matrix should be converted into a vector for easier comparison with other GatingHierarchy objects.}
}
\value{
A binning matrix and (optionally) its vectorized form.
}
\description{
Binning flow cytometric data
}
